scanner2: rename package scanner2 -> scanner
