+++ lexer/lexer.go
@@ -3,7 +3,7 @@ package lexer
 import (
 	"io"
 
-	scanner "github.com/uji/solparser/scanner2"
+	"github.com/uji/solparser/scanner"
 	"github.com/uji/solparser/token"
 )
