+++ scanner/scanner.go
