+++ scanner/scanner_test.go
